requirements.txt
django
gunicorn
django-on-heroku
djangorestframework
python-dotenv